# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Human Time Market API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rfps:
    get:
      operationId: listRfps
      tags: [rfps]
      summary: Browse open RFPs
      parameters:
        - name: skillCategoryId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/RfpStatus"
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated list of RFPs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RfpsPage"
    post:
      operationId: createRfp
      tags: [rfps]
      summary: Post a new RFP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRfpBody"
      responses:
        "201":
          description: Created RFP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RfpDetail"
  /rfps/{rfpId}:
    get:
      operationId: getRfp
      tags: [rfps]
      summary: Get RFP detail with responses
      parameters:
        - name: rfpId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: RFP detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RfpDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rfps/{rfpId}/responses:
    post:
      operationId: respondToRfp
      tags: [rfps]
      summary: Submit a response to an RFP
      parameters:
        - name: rfpId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRfpResponseBody"
      responses:
        "201":
          description: Response submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RfpResponseDetail"
        "400":
          description: Cannot respond
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RfpStatus:
      type: string
      enum: [open, closed, fulfilled]
    RfpResponseStatus:
      type: string
      enum: [submitted, accepted, rejected, withdrawn]
...
    RfpSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        skillCategoryId:
          type: integer
        skillCategoryName:
          type: string
        skillCategoryParentName:
          type: ["string", "null"]
        budgetMinCents:
          type: integer
        budgetMaxCents:
          type: integer
        hoursNeeded:
          type: integer
        deadline:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/RfpStatus"
        buyerId:
          type: integer
        buyerDisplayName:
          type: string
        responseCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - skillCategoryId
        - skillCategoryName
        - budgetMinCents
        - budgetMaxCents
        - hoursNeeded
        - deadline
        - status
        - buyerId
        - buyerDisplayName
        - responseCount
        - createdAt
    RfpDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        skillCategoryId:
          type: integer
        skillCategoryName:
          type: string
        skillCategoryParentName:
          type: ["string", "null"]
        budgetMinCents:
          type: integer
        budgetMaxCents:
          type: integer
        hoursNeeded:
          type: integer
        deadline:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/RfpStatus"
        buyerId:
          type: integer
        buyerClerkId:
          type: string
        buyerDisplayName:
          type: string
        responses:
          type: array
          items:
            $ref: "#/components/schemas/RfpResponseDetail"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - skillCategoryId
        - skillCategoryName
        - budgetMinCents
        - budgetMaxCents
        - hoursNeeded
        - deadline
        - status
        - buyerId
        - buyerClerkId
        - buyerDisplayName
        - responses
        - createdAt
        - updatedAt
    RfpsPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RfpSummary"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - items
        - total
        - limit
        - offset
    CreateRfpBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        skillCategoryId:
          type: integer
        budgetMinCents:
          type: integer
        budgetMaxCents:
          type: integer
        hoursNeeded:
          type: integer
        deadline:
          type: string
          format: date
      required:
        - title
        - description
        - skillCategoryId
        - budgetMinCents
        - budgetMaxCents
        - hoursNeeded
        - deadline
    RfpResponseDetail:
      type: object
      properties:
        id:
          type: integer
        rfpId:
          type: integer
        professionalId:
          type: integer
        professionalDisplayName:
          type: string
        professionalClerkId:
          type: string
        proposedRateCents:
          type: integer
        message:
          type: string
        status:
          $ref: "#/components/schemas/RfpResponseStatus"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - rfpId
        - professionalId
        - professionalDisplayName
        - professionalClerkId
        - proposedRateCents
        - message
        - status
        - createdAt
    CreateRfpResponseBody:
      type: object
      properties:
        proposedRateCents:
          type: integer
        message:
          type: string
      required:
        - proposedRateCents
        - message
...